fix(ui): widen node-index storage past 255 (uint16_t count/parent/subtreeEnd)
Closes the 255-node ceiling — the last lower-priority item from the index-type
audit. __ui_node_count was uint8_t (max 255), parent/subtreeEnd were uint8_t
with UI_NO_PARENT=255 colliding with a valid index, and ~30 loop counters/locals
walking the node table were uint8_t and would wrap at 255.

Widened to uint16_t: __ui_node_count (decl + emitter), the parent/subtreeEnd
node fields, the six binding/transition/animation struct node fields, all 13
node-count-bounded loops, the subtreeEnd-bounded loops, the node-index
params/locals (nodeIdx/scrollNode/animNode/clickedNode + parent walkers), and
removed ~30 truncating (uint8_t) casts on node-index expressions. Bumped
UI_NO_PARENT 255 → 0xFFFF. The lowering emits 0xFFFF for no-parent.

The preview runtime uses JS numbers, so it needed no changes. The demo (144
nodes) is well under the new 65535 ceiling. Regression test asserts uint16_t
count/parent/subtreeEnd, the 0xFFFF sentinel, and uint16_t loop counters.

+++ packages/cuttlefish/src/emit/emitters/ui-emitter.ts
@@ -203,7 +203,7 @@ export function emitUIRuntime(ctx: EmitterContext): void {
     (sum, { lowered }) => sum + countNodes(lowered.nodeTable),
     0,
   );
-  ctx.sourceLines.push(`const uint8_t __ui_node_count = ${totalNodes};`);
+  ctx.sourceLines.push(`const uint16_t __ui_node_count = ${totalNodes};`);
   ctx.sourceLines.push(`const uint8_t __ui_trans_count = ${countTransitions()};`);
   ctx.sourceLines.push(`const uint8_t __ui_binding_count = ${uiBindings().length};`);
   // Count screens across all modules (for multi-screen navigation).
